Add k8s readiness check
This change adds a readiness check to allow for the provisioning
server to properly wait for the cluster to be prepared for running
workloads.

To accomplish this, the ability to run commands silently was also
added so that we can run kops validation checks on a regular
basis without spamming logs.