Add k8s client and deploy mattermost-operator
This change adds a kubernetes client to the provisioner. This
client is able to manipulate and create the mattermost custom
resources that we use for mattermost cluster installations.

The API objects for these custom resources, as well as the
documentation for generating the code from these objects, can
be found in `internal/tools/k8s`.

Also, the cluster creation process now deploys the mattermost
operator. This leverages the new kubernetes client to configure
the necessary cluster objects.