MM-14996: REST API proof-of-concept (#10)
* MM-14996: REST API proof-of-concept

Introduce the REST API, rewriting the existing cli to leverage same (and require a `--server` parameter). Implement the `server` command to start a provisioning server to expose the API. Introduce a Go client for the API to facilitate unit-testing (currently remaining `internal`).

Note that there are a few new TODOs, largely around the pending changes to make this API asynchronous. Rather than jumping through hoops with the current model, I've just noted where changes need to be made.

* better invalid argument handling

* fix default server invocation

* leverage subrouters and regex