Use mattermost clientset from mattermost-operator
The mattermost-operator repo is now the single source of truth
for the `ClusterInstallation` k8s API. This change imports the
code generated there for future manipulation of the mattermost
k8s resources.

This change also updates the k8s manifests for deploying
operators to the newest versions found in mattermost-operator.