introduce provisioner.PrepareCluster step
This breaks out the part of the provisioner that needs to modify the
cluster (to record changes to the ProvisionerMetadata) from the part
that actually provisions, making it explicit in the supervisor that we
need to write said changes back to disk.

I had incorrectly removed the "save-after-provision" step when an error
occurs during the provisioning. Hopefully this will make it more
obvious.