Add Mattermost config get/set functionality
This initial implementation is done by executing the Mattermost
CLI in one of the running containers of the cluster installation.

In the future, we should move to a direct API approach and should
use Kubernetes jobs. That said, this change lays the groundwork
for running arbitrary commands in the Mattermost pods.