Add separate cluster 'provisioning' state
This new state contains the logic for creating the k8s operators
that are necessary to deploy Mattermost installations. This process
used to occur directly after cluster creation with no separation.
Now, clusters can be 'provisioned' again in the stable state, which
will update all operator resources to the newest version contained
in the provisioner manifests.

This change also does the following:
 - Refactors much of the k8s client library to support updates of
   previously-created k8s resources.
 - Fixes a unit test bug where we were not properly checking that
   the right cluster states were being returned for work by the
   supervisor.
 - Cleans up logging in various places.