Add cloud server webhooks (#68)
This change introduces webhooks to the cloud provisioning server.
These webhooks can be registered in order to receive a stream of
events from the server. The payload of the webhooks contains state
information the various clusters, installations, and cluster
installations as the server moves them from one state to another.
This allows clients of the cloud server to know when resources
they care about have reached a desired state or for general logging
and troubleshooting.