Flexible server supervisor options
This change introduces options for controlling which supervisors
run on server startup. This allows for control over the work the
server will do on clusters, installations, and cluster installations.

Due to installation and cluster installation work requiring no
persistent cluster data, we can now run cloud servers in HA as long
as the cluster supervisor is not started.

Finally, this allows for running the cloud server in an API-only
mode by disabling all supervisors.