MM-21259 Tag VPCs with Human Identifiable Name (#134)
Use the new CLOUD_SERVER_OWNER environment variable to have the provisioner server tag the VPC with the value provided, or do nothing and the server will attempt to tag the VPC with the configured git email from the local environment. If all that fails, the VPC will be tagged with "SRETeam".