Allow unset with an empty MattermostEnv
A nil pointer stored in p.MattermostEnv will cause no change.

An empty map stored in p.MattermostEnv will cause all environment
variables to unset.

A map with values stored in it, stored in p.MattermostEnv, will
replace all environment variables on the target with the ones in
p.MattermostEnv.