Add group supervisor for installation updates (#176)
The group supervisor is responsible for managing rolling updates
to installations when a group's configuration is updated. It does
this by moving installations in the group to an update-requested
state while respecting the group's MaxRolling setting.

Note that the cloud server still does not apply merged group config
to installations so the group supervisor is disabled by default.
This will be changed in a future commit.

This change also corrects a SQL query bug as well as a bug with
how group merged config overrides are reported.