Add option to change cluster AMI on upgrade
This new functionality allows the cloud provisioner to update the
various instance groups in a cluster to use a new AMI for the nodes.
Upgrades can now handle an update to the kubernetes version, AMI,
or both simultaneously. Passing in an empty value for the AMI will
use the default kops AMI for that kubernetes version.