Stale installation supervisor fix
A devious bug has been uncovered where supervisors can act on
installation data that has become "stale" due to other cloud servers
working on the installation beforehand.

This flow can occur when there are a high amount of installations
that require work and the multiple cloud servers are working on
the same tasks. They could end up getting a slice of installations
that need work at the same time and then processing those installations
independently over time with the outdated information regarding state.

This change forces the installation supervisor to check state on
these installations and abort if the state is outdated. When the
supervisor wakes up later it will get a new set of work and will
proceed.

This also cleans up some multitenant database logging.