Graceful SIGTERM shutdown
Kubernetes has a process for terminating pods as part of rolling
updates and other pod lifecycle mechanisms. One of the steps of
this process is sending a SIGTERM signal to the containers in the
pod.

The provisioner previously was not listening for SIGTERM properly,
meaning that there was a chance that supervisors could be performing
work when the container was terminated. This change handles SIGTERM
properly by attempting a graceful shutdown of the cloud server.

Also, logging has been added with a Shutdown() method for each
supervisor running in the server.