Refactor cluster metadata handling (#236)
This change alters the behavior of accessing cluster metadata.
Before, it was presented in the cluster object in raw format which
required marshaling and unmarshaling to access and update. This
logic is now handled directly on the store level so the returned
clusters now have metadata in their actual types. When cluster
updates are made, the store will marshal these types back into
their raw format before being stored.