Add cluster workbench command for cloud client
This new command can be used to create a local temporary workbench
directory for a cluster. This directory contains the kubecfg and
terraform files for that cluster. This flow replicates the process
that the provisioning server uses to manage clusters, which can be
useful for troubleshooting issues or manual cluster updates.