Add cluster scale up logic (#259)
This adds logic for scaling up cluster worker nodes when scheduling
new installations. If the cloud server is configured to do so, worker
nodes will be increased up to the cluster worker node maximum if the
new installation would exceed the cluster resource threshold. The
cluster scaling and installation reconciliation will happen in
parallel.