Add cloud resource API security locking
This introduces the concept of cloud resource API security locking
for clusters, installations, cluster installations and groups. When
a resource is locked, no changes can be made to the resource until
it is unlocked again.

Locking and unlocking occurs via a new set of cloud server API
endpoints under /security. By separating the ability to control
resource locks out from the normal endpoints, the security endpoints
can be blocked off from other consumers of the cloud server API.
In the future, these endponts could also be exposed on a separate
port from the normal API if desired.

When creating new cloud resources, it is now possible to create them
in a state where they are immediately locked from API changes. This
opens up the possibility for clients of the cloud server API to
create resources, but not modify or delete which offers greater
control over their basic permissions. This functionality should
be especially helpful when dealing with long-lived resources that
are not changed often.