Enable default AWS S3 SSE
This sets the default encryption scheme for new installation S3
buckets to use AES-256 encryption. The Mattermost filestore SSE
configuration option is also set to true to force the behavior
when a license is present.

This also corrects an issue where the cloud installation ID env
var was lost when updating installations.