Handle multitenant filestore teardown with no cluster installations
It is possible that quickly deleting a new installation before it
has created cluster installation resources will result in a teardown
failure if the installation had a multitenant filestore type. This
change checks for an absence of cluster installations and will
assume teardown is complete if there are not any.