Add group no-op update sequence bump option (#332)
The group update API now allows for a group's sequence number to
be increased by one with no changes to the group. This allows for
a clean way to ensure all installations in a group have the latest
baseline configuration and resources without making any changes to
the group.

This change also includes a change where GroupID and Sequence
number will be added as manifest labels on the ClusterInstallation
custom resource. These labels can be used for more efficient
troubleshooting.