Add bifrost filestore option
Bifrost is a new filestore option for installations which acts as
a shared AWS S3 bucket gateway. Bifrost is deployed and managed as
a shared cluster utility via the provisioning process.

The service relies on a shared S3 bucket being created in each
cluster VPC as well as IAM credentials for accessing that bucket
being available via AWS secret manager.

This change also includes AWS client refactoring and logging cleanup.