Replace object reassignment with value assignments
This was creating a bug caught by the tests (woohoo!)
I thought these were equivalent statements but I guess I'm not too
surprised to see that I was incorrect about that.
Lesson here: don't reassign the receiver to a new pointer from inside
of a method. I was worried that wouldn't work, originally. Lesson 2 is
to listen to your gut when it tells you something isn't right.