Installation database weighting (#426)
This change introduces the concept of database weights to
installations. Installations that are hibernating now reserve less
space in a shared database compared to other states. This allows
for the spare capacity of multiple hibernating installations to be
allocated to new installations while still offering some protection
if these installations wake up.