Add installation recovery command (#456)
This new command uses direct store and AWS client connections to
update deleted installation resources so that the provisioning
server will recreate them. Currently, only certain database and
filestore types are supported. In the future, this logic will be
moved directly into the provisioning server and exposed via an API.