Create k8s license secrets with unique names (#457)
This change results in the provisioner creating license secrets
with unique names based on the installation license secret value.
This allows the operator to know that the mattermost app pods
should be rolled to obtain this new value. This also ensures that
external systems can properly wait for license updates to occur
via provisioner webhooks.