Add PGBouncer installation databases (#472)
This new database type pools database connections via the PGBouncer
utility running inside the cluster. Installations are assigned to
a logical database in the database cluster and are given a schema
in which to store their data. Authentication is performed through
an authorization query from a pgbouncer user to the backing database
to ensure the user credentials are valid.

This change also includes a large amount of refactoring in the AWS
package in an attempt to reuse existing logic.