Implement first SLI for installations.
When this commit applied every new installation will have it's first SLI for API request errors.
On hibernation and deletion events the SLI will get removed, while on wake-up will get recreated.
Also if an installation doesn't have SLI the deletion will not fail because of this and in any cluster installation update event SLI is going to be created if it not exists.
On top of this SLI it will be built SLO alerting and monitoring for every installation
Issue: MM-37305