Add a new tool for testing provisioner installations like a real user (#489)
The tool is called testwick and can be build by itself as a CLI tool. The
implementation though is like a framework which you can use by itself and
write your own scenarios but of course always respecting the order of the
actions.

Right now it creates installations and tear down them in the end. It runs
couple of actions to these installations like:
- Create Channel
- Create Users
- Create Teams
- Add team members
- Post messages

Ticket: https://mattermost.atlassian.net/browse/MM-37408