Change k8s resource calculations (#569)
Previously, resource calculations done on k8s clusters would take
all pods and nodes into account. This is an issue as custom kops
instance groups and master nodes could not run standard workloads
and should be ignored.

This change alters the calculation to only check the standard
worker nodes and the pods that are running on those nodes.