Add velero in our kops cluster
We are adding velero as a utility tool for start setting up the disaster recovery plan
goals. In case of a k8s cluster cannot be fully functional for any reason we have a backup
which we can use to create a k8s cluster with the last backup.

Ticket: https://mattermost.atlassian.net/browse/MM-43762