Add installation deletion queuing (#700)
This new logic changes how installations are deleted by the
provisioner. Now, installations enter a deletion pending state
when they were previously in a state that enabled them to have
created persistent data. This deletion pending state acts as a
type of 'soft deletion' where actual deletion of the installation
and its data occurs at a later time as configured in the provisioner.
This time can be lowered to zero to provide immediate deletion
similar to the existing behavior.