Add cluster-installation `ready` state
The Mattermost Operator exposes a ready state which indicates that
an installation is nearly stable and can handle incoming requests.
Instead of always translating this state to stable internally, this
change records the ready state so that actions can be taken if desired
with context that the installation is not completely stable yet.

The installation supervisor will now proceed with creation when
an installation is ready, while other flows such as installation
updates require the installation to be stable before proceeding.