Add external database type support
This new database type can be used to create installations with
any precreated database as long as the installation can successfully
connect. The credentials for external databases are pulled down
from a precreated AWS Secret Manager secret. No management of the
database or the secret are performed by the provisioner.