Add external database type support (#729)
This new database type can be used to create installations with
any pre-created database as long as the installation can successfully
connect. The credentials for external databases are pulled down
from a pre-created AWS Secret Manager secret. No management of the
database or the secret are performed by the provisioner.