[HotFix] Route53 DNS RecordSet retrieval (#835)
* Migrate from CircleCI to Github Actions (#790)

We gradually moving away from CirleCI to Github Actions as we are consolidating
the CIs we use. In parallel we are introducing `trivy` in the PR level
and we block the PR about the number of critical vulnerabilities.

Ticket: https://mattermost.atlassian.net/browse/CLD-4697

* Fix latest tag docker push and ignore CD on master (#832)

We removed the latest tag during the migration and now we return it back for docker images. In
parallel we ignore CD during master branch.

Ticket: https://mattermost.atlassian.net/browse/CLD-4840

* [MM-49661] Fixed retrieval of DNS Record Sets (#834)

* fixed exportloopref

* added extra log fields

* Tag docker images with latest tag (#833)

* disable rotation on cluster resize (#836)

Co-authored-by: Spiros Economakis <812075+spirosoik@users.noreply.github.com>
Co-authored-by: Gabe Jackson <gabe@coffeepowered.co>