Support arbitrary domain creation for installations (#983)
We were previously blocking all domain additions for installations
that didn't have a subdomain that matched the `name` value. There
were many reasons for this, but it was inflexible and needed to be
improved. There is more work to do in supported custom domains, but
this allows for completely switching an installation's domain after
creation.

This change does the following:
 - Allows for creating any new installation domain assuming the
   provisioner has access to the zone.
 - Allows for deleting domains for an installation after they
   are no longer marked as primary.
 - Fixes a bug where certain domains couldn't be selected as
   primary.
 - Improves domain management logging.