Add logic to skip AMI updates for instance groups
There are situations where it is ideal to not apply the standard
AMI to all kops instance groups in a cluster. This change allows
for setting an instance group metadata label with a key of
`mattermost/cloud-provisioner-ami-skip` and any value to instruct
the provisioner to skip any AMI updates on that instance group.