Separate installation scheduling locks
This change introduces a new type of cluster lock that is used for
installation scheduling. This lock is used to coordinate new
installation creation without requiring a full lock on the
underlying cluster. This allows for one provisioner instance to
update the cluster while another instance can schedule new
installations assuming the cluster is in a valid state.

This allows for scheduling installations on clusters that are
being provisioned, resized, or upgraded.