Add Mattermost debug data gathering
This is initial work to allow for obtaining pprof information from
a given Mattermost deployment. The heap and goroutine performance
profiles can now be obtained in a zip file for all pods in a
cluster installation. The zip is processed on the server and then
sent to the cloud client to be reviewed.