Add installation deletion scheduling
This change adds optional installation deletion scheduling logic
to the provisioner. Installations can now be created with a
configurable TTL which can useful for quick testing and other
use-cases.

Installations with scheduled deletion will only be deleted if they
are not deletion-locked and the current time is past their scheduled
deletion time. Deletion-locked installations will be deleted once
their lock is removed if the scheduled deletion time has passed.

This is completely separate from the existing deleting pending
logic which allows for specifying how long an installation remains
in a deletion-pending state after it was scheduled for deletion.