Remove Argo CD utility logic
The Argo CD utility logic was used before we migrated to EKS clusters
and is now unused and untested. Due to conflicts with trying to upgrade
other k8s go modules, we are opting to remove the Argo dependencies and
clean up the codebase instead of trying to update Argo.