Add local-ephemeral filestore option
This is the initial implementation of a new local ephemeral filestore
backend option for installations. This new option is intended to be used
by very short-lived installations such a trials. The operator doesn't
directly support this type of backend so a combination of S3
configuration and env overrides are used to instruct Mattermost to use a
local directory for filestore data.