Add PodDisruptionBudget policy for HA Mattermost installations
When an installation has a pod replica count of at least 2 we now
create a PodDisruptionBudget policy for the pod with a minimum available
value of 1. This ensures better uptime guarantees for these
installations.

This change also corrects an issue with some instances when group
scheduling values are not set.